rip out bamboohr
